Make StateVariable.cc compile with GNU C++ 3.4.6 without warnings.
git-svn-id: https://gxemul.svn.sourceforge.net/svnroot/gxemul/gxemul/trunk@23 dbbdf1f2-f3a0-11dc-8807-f0e41a653009

+++ src/main/StateVariable.cc
@@ -684,7 +684,7 @@ static void Test_StateVariable_Numeric_SetValue()
 	UnitTest::Assert("changing to -110 should be possible",
 	    vsint8.SetValue("-110") == true);
 	UnitTest::Assert("varSInt8 should have been updated",
-	    varSInt8, -110);
+	    varSInt8, (uint64_t) -110);
 
 	// Tests for other numeric types: TODO
 }
